+++ NEWS.md
@@ -1,3 +1,11 @@
+# vici 0.6.0
+
+ * Added support for additional degree of freedom methods 
+
+# vici 0.5.3
+
+ * Added source code & CRAN link buttons
+
 # vici 0.5.1
 
  * First completely operational version of the app